refactor: remove redundant course type
