+++ packages/zod-schemas/courses.response.schema.ts
@@ -9,7 +9,6 @@ import {
   TIME_REGEX,
   vote,
 } from "./constants.js";
-import { CourseInfoSchema } from "./courses.schema.js";
 
 export const ClassSchema = z.object({
   id: z.string(),
@@ -35,22 +34,50 @@ export const SectionSchema = z.object({
   classes: z.array(ClassSchema),
 });
 
-export const CourseDetailSchema = z.object({
+// 1. Sub-schema for the 'course' object
+export const CourseSchema = z.object({
   id: z.string(),
   studyProgram: studyProgram,
   academicYear: z.number().int(),
   semester: semester,
   courseNo: z.string(),
-  courseCondition: z.string().nullable(),
+  courseCondition: z.string().nullish().default("-"),
+  genEdType: genEdType,
   midtermStart: z.string().datetime().nullable(),
   midtermEnd: z.string().datetime().nullable(),
   finalStart: z.string().datetime().nullable(),
   finalEnd: z.string().datetime().nullable(),
-  genEdType: genEdType,
-  createdAt: z.string().datetime().optional(),
-  updatedAt: z.string().datetime().optional(),
-  courseInfo: CourseInfoSchema,
   sections: z.array(SectionSchema),
+});
+
+// 2. Sub-schema for the 'courseInfo' object
+export const CourseInfoSchema = z.object({
+  abbrName: z.string(),
+  courseNameEn: z.string(),
+  courseNameTh: z.string(),
+  courseDescEn: z.string().nullable(),
+  courseDescTh: z.string().nullable(),
+  faculty: z.string().nullable(),
+  department: z.string().nullable(),
+  credit: z.string(),
+  creditHours: z.string().nullable(),
+});
+
+// 3. Sub-schema for the 'stats' object
+const StatsSchema = z.object({
+  sectionsCount: z.number().int(),
+  capacitySum: z.number().int(),
+  remainingSum: z.number().int(),
+  hasSeats: z.boolean(),
+  isClosedAll: z.boolean(),
+});
+
+// 4. The Main Schema combining everything
+export const CourseDetailsSchema = z.object({
+  course: CourseSchema,
+  courseInfo: CourseInfoSchema,
+  stats: StatsSchema,
+  fitMySchedule: z.boolean(),
 });
 
 export const CourseReview = z.object({
@@ -69,7 +96,9 @@ export const CourseReview = z.object({
 });
 
 export const CourseNoResponseSchema = z.object({
-  course: CourseDetailSchema,
+  course: CourseSchema.extend({
+    courseInfo: CourseInfoSchema,
+  }),
   reviews: z.array(CourseReview),
 });
